chore: Moved stack-trace-parser.service.mts from Service/Logger/StackTraceParser to Service/stack-trace-parser.
