+++ packages/architectura/src/Service/Logger/index.mts
@@ -1,3 +1,3 @@
+export * from "../stack-trace-parser/stack-trace-parser.service.mjs";
 export * from "./LogLevelEnum.mjs";
-export * from "./stack-trace-parser.service.mjs";
 export * from "./StackTraceParser/StackTraceParserLineAndPositionInterface.mjs";
