+++ packages/architectura/src/Service/logger.service.mts
@@ -2,7 +2,7 @@ import { Time } from "../Core/Time.mjs";
 
 import { LogLevelEnum } from "./Logger/LogLevelEnum.mjs";
 
-import { StackTraceParser } from "./Logger/stack-trace-parser.service.mjs";
+import { StackTraceParser } from "./stack-trace-parser/stack-trace-parser.service.mjs";
 
 class LoggerService
 {
